Improve suggest term link (#271)
* Add missing search term to suggestion link

Having a GitHub link to directly suggest a term is great, and it
could be even better if we could pass on the term the user already
typed onto the address bar. Link magic!

* Highlight the missing term

Now that we have access to it, we could show it on the page and
highlight it a bit. Otherwise, it is only on the address bar. Added
“yet” as well as a small hint that this glossary grows and people’s
suggestions are valuable.

* Cleanup GitHub’s new issue link

A couple of parameters we were passing were not being used (no
content present), so we can provide a simpler user experience by
removing them for the time being.